pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ura-ds-portfolio"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pandas>=2.0",
  "numpy>=1.24",
  "scikit-learn>=1.3",
  "xgboost>=2.0",
  "shap>=0.44",
  "pyyaml>=6.0",
  "jsonschema>=4.19",
  "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "ruff>=0.5",
]

[tool.setuptools.package-dir]
"" = "projects/wht-exemption/src"

[tool.setuptools.packages.find]
where = ["projects/wht-exemption/src"]

[tool.ruff]
line-length = 100
